+++ src/history.h
@@ -3,5 +3,5 @@
 #include "movegen.h"
 
 void update_histories(Position* pos, int depth, uint16_t* moves, int length);
-int16_t get_history(Position* pos, uint16_t move);
+int32_t get_history(Position* pos, uint16_t move);
 #endif
